+++ yospace/src/main/java/com/bitmovin/player/integration/yospace/CustomEvent.kt
@@ -12,7 +12,7 @@ open class CustomEvent {
     /**
      * The time at which the event was emitted in milliseconds since the Unix Epoch.
      */
-    var timestamp: Long = System.currentTimeMillis()
+    var timestamp: Long = System.currentTimeMillis() / 1000
 }
 /**
  * Includes all possible events that the [Source] can emit.